+++ internal/api/web/views/layout.templ
@@ -1,6 +1,7 @@
 package views
 
 templ Layout() {
+	<!DOCTYPE html>
 	<html lang="en">
 		<head>
 			<meta charset="utf-8"/>
